:sparkles: :: add responsive article-list component
