+++ src/app/blog/styles.module.css
@@ -45,27 +45,6 @@
   line-height: normal;
 }
 
-.postrow__upper {
-  position: relative;
-  width: 100%;
-  flex-shrink: 0;
-
-  display: flex;
-  justify-content: space-between;
-}
-
-.postrow__lower {
-  position: relative;
-  width: 100%;
-  flex-shrink: 0;
-
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-auto-rows: auto;
-  grid-column-gap: 1.5rem;
-  grid-row-gap: 3rem;
-}
-
 .centered__row {
   position: relative;
   width: 100%;
