+++ src/app/components/article-list/articlelist.module.css
@@ -1,0 +1,31 @@
+.postrow {
+  position: relative;
+  width: 100%;
+  flex-shrink: 0;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  row-gap: 1.5rem;
+}
+
+.postrow__upper {
+  position: relative;
+  width: 100%;
+  flex-shrink: 0;
+
+  display: flex;
+  justify-content: space-between;
+}
+
+.postrow__lower {
+  position: relative;
+  width: 100%;
+  flex-shrink: 0;
+
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-auto-rows: auto;
+  grid-column-gap: 1.5rem;
+  grid-row-gap: 3rem;
+}
